+++ lib/emque/consuming/configuration.rb
@@ -3,23 +3,29 @@ require "logger"
 module Emque
   module Consuming
     class Configuration
-      attr_accessor :app_name, :auto_shutdown, :adapter, :error_handlers,
-        :error_limit, :error_expiration, :status, :status_port, :status_host,
+      attr_accessor :app_name, :adapter, :auto_shutdown,
+        :delayed_message_workers, :enable_delayed_message, :error_handlers,
+        :error_limit, :error_expiration, :retryable_errors,
+        :retryable_error_limit, :status, :status_port, :status_host,
         :socket_path, :shutdown_handlers
       attr_writer :env, :log_level
 
       def initialize
-        @app_name           = ""
-        @auto_shutdown      = false
-        @error_handlers     = []
-        @error_limit        = 5
-        @error_expiration   = 3600 # 60 minutes
-        @log_level          = nil
-        @status_port        = 10000
-        @status_host        = "0.0.0.0"
-        @status             = :off # :on
-        @socket_path        = "tmp/emque.sock"
-        @shutdown_handlers  = []
+        @app_name                = ""
+        @auto_shutdown           = false
+        @delayed_message_workers = 1
+        @enable_delayed_message  = false
+        @error_handlers          = []
+        @error_limit             = 5
+        @error_expiration        = 3600 # 60 minutes
+        @log_level               = nil
+        @retryable_errors        = []
+        @retryable_error_limit   = 3
+        @status_port             = 10000
+        @status_host             = "0.0.0.0"
+        @status                  = :off # :on
+        @socket_path             = "tmp/emque.sock"
+        @shutdown_handlers       = []
       end
 
       def env
@@ -42,13 +48,17 @@ module Emque
         {}.tap { |config|
           [
             :app_name,
-            :auto_shutdown,
             :adapter,
+            :auto_shutdown,
+            :delayed_message_workers,
             :env,
+            :enable_delayed_message,
             :error_handlers,
             :error_limit,
             :error_expiration,
             :log_level,
+            :retryable_errors,
+            :retryable_error_limit,
             :status_port,
             :status_host,
             :status,