+++ lib/emque/consuming/consumer.rb
@@ -1,5 +1,6 @@
 require "oj"
 require_relative "consumer/common"
+require_relative "retryable_errors"
 
 module Emque
   module Consuming