+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Emque Consuming CHANGELOG
 
+- [Add in the ability to retry errors and back off with an exponential delay](https://github.com/emque/emque-consuming/pull/55) 1.2.0
 - [Add in a configuration option to disable auto shutdown on reaching the error limit](https://github.com/emque/emque-consuming/pull/58) 1.1.3
 
 ## 1.0.0.beta4
@@ -10,6 +11,6 @@ After starting up, messages will need to be manually moved
 from the old queue to the new one.
 
 ### Failed Message Routing
-Messages that are not acknolwedged due to a consumer error will now be routed
+Messages that are not acknowledged due to a consumer error will now be routed
 into a `service_name.error` queue. These can then be inspected and be discarded,
 purged, or manually moved back to the primary queue for re-processing.